Rename directory to match the package name
