+++ pkg/clusteragent/admission/mutate/autoinstrumentation/libraryinjection/init_container.go
