+++ pkg/clusteragent/admission/mutate/autoinstrumentation/libraryinjection/init_container_test.go
